Disabled stty hiding
`read` is blocking Ctrl-C interrupt, so that means `kpm` needs to use a
Kill to properly kill the script execution. Because kpm is using a kill,
the scripts don't get a chance to set the stty echo back on, causing the
echo to be hidden.

So, we're disabling stty -echo for the moment, until this whole
situation can be resolved in a way that won't harm user experience.