Restart mongo after mongodb.conf change
The mongodb restart was executing "sudo service mongodb start" which only works if the mongodb service is not already running. As the final bit of the mongodb install actually starts the service, this part fails, stops the script, and prints errors. I modified "start" to "restart" which alleviates the issue.